src/requirements.txt
Flask==2.2.3
Jinja2==3.1.2
markupsafe==2.1.2
lxml==4.9.2
requests==2.28.2
appengine-python-standard==1.1.1
python-dateutil==2.8.2
werkzeug==2.2.3
six==1.16.0
#IPy==1.01								# カスタマイズして使用・・・「222.92.143.86/24」などがInvalid扱いになってしまうので
pycryptodome==3.17
pkcs7==0.1.2
google-cloud-logging==3.5.0
#asn1crypto==0.24.0
#beautifulsoup4==4.8.0
#biplist==1.0.3
#certifi==2019.9.11
#cffi==1.12.3
#chardet==3.0.4
#Click==7.0
#cryptography==2.7
#idna==2.8
#itsdangerous==1.1.0
#MarkupSafe==1.1.1
#pycparser==2.19
#pyOpenSSL==23.1.1
#soupsieve==1.9.3
#firebase-admin
#google-cloud-firestore
#flask_session==0.4.0		# カスタマイズして使用
#flask-gae-sessions				# カスタマイズして使用
google-api-python-client==2.80.0
apiclient==1.0.4
gspread==5.7.2
pyyaml==6.0
#ssl
sseclient-py==1.7.2
Flask-Cors==3.0.10
firebase-admin==6.1.0
numpy==1.23.5
redis==4.5.4
pyquery==2.0.0
python-magic==0.4.27
openai==0.27.6
tenacity==8.2.2
tiktoken==0.3.3
spacy==3.5.2
fake-useragent==1.1.3
parse==1.19.0
w3lib==2.1.1
# メモリ使用量が大きいライブラリ.現状未使用なのでいったんコメントアウト
#python-docx==0.8.11
#python-dotenv==1.0.0
#python-pptx==0.6.21
#PyPDF2==3.0.1
#beautifulsoup4==4.12.2
#pyppeteer==1.0.2
elasticsearch>=7.17.9,<8.0.0
opensearch-py>=2.2.0
#certifi==2023.5.7
#chardet==5.1.0
urllib3
ipaddress
bcrypt==4.0.1
gunicorn==20.1		# GAEGEN2対応：メモリ使用量チューニング
psutil==5.9.5				# GAEGEN2対応：メモリ使用量チューニング
